refactor(github): extract GitHubError into errors module
Move the GitHubError enum and its classify() method out of github.rs (467)
into a new github/errors.rs (105), re-exported as github::GitHubError so the
GitHubClient impl and external callers are unchanged. classify_tests moves to
github/errors/ alongside its subject (it exercises only GitHubError::classify,
not GitHubClient). github.rs drops to 368 lines; production behavior unchanged;
449 tests pass.
